fix edit and delete for question

+++ src/app/modules/coffee-lab/components/forum-menu/forum-menu.component.ts
@@ -4,6 +4,8 @@ import { CoffeeLabService, GlobalsService } from '@services';
 import { environment } from '@env/environment';
 import { ToastrService } from 'ngx-toastr';
 import { Router } from '@angular/router';
+import { DialogService } from 'primeng/dynamicdialog';
+import { ConfirmComponent } from '@shared';
 
 @Component({
     selector: 'app-forum-menu',
@@ -27,6 +29,7 @@ export class ForumMenuComponent implements OnInit {
         private coffeeLabService: CoffeeLabService,
         private toastService: ToastrService,
         private router: Router,
+        private dialogService: DialogService,
     ) {}
 
     ngOnInit(): void {
@@ -137,17 +140,17 @@ export class ForumMenuComponent implements OnInit {
     onEdit(): void {
         switch (this.forumType) {
             case 'question':
-                this.router.navigate(['/coffee-lab/create-post/tab/question'], {
+                this.router.navigate(['/coffee-lab/create-post/tab'], {
                     queryParams: { id: this.selectedItem.id },
                 });
                 break;
             case 'article':
-                this.router.navigate(['/coffee-lab/create-post/tab/article'], {
+                this.router.navigate(['/coffee-lab/create-post/tab'], {
                     queryParams: { id: this.selectedItem.id },
                 });
                 break;
             case 'recipe':
-                this.router.navigate(['/coffee-lab/create-post/tab/recipe'], {
+                this.router.navigate(['/coffee-lab/create-post/tab'], {
                     queryParams: { id: this.selectedItem.id },
                 });
                 break;
@@ -157,13 +160,26 @@ export class ForumMenuComponent implements OnInit {
                 break;
         }
     }
+
     onDelete(): void {
-        this.coffeeLabService.deleteForumById(this.forumType, this.selectedItem.id).subscribe((res: any) => {
-            if (res.success) {
-                this.toastService.success(`You have deleted a ${this.forumType} successfully.`);
-                this.coffeeLabService.forumDeleteEvent.emit();
-            } else {
-                this.toastService.error(`Failed to delete a ${this.forumType}.`);
+        this.dialogService
+            .open(ConfirmComponent, {
+                data: {
+                    type: 'delete',
+                },
+                showHeader: false,
+                styleClass: 'confirm-dialog',
+            })
+            .onClose.subscribe((action: any) => {
+            if (action === 'yes') {
+                this.coffeeLabService.deleteForumById(this.forumType, this.selectedItem.id).subscribe((res: any) => {
+                    if (res.success) {
+                        this.toastService.success(`You have deleted a ${this.forumType} successfully.`);
+                        this.coffeeLabService.forumDeleteEvent.emit();
+                    } else {
+                        this.toastService.error(`Failed to delete a ${this.forumType}.`);
+                    }
+                });
             }
         });
     }
